[develop] 将读取数据换成BufferedReader

+++ src/main/java/cn/edu/nju/lc/data/CollectData.java
@@ -6,7 +6,7 @@ public class CollectData {
      * {@link DataType}
      */
     private int type;
-    private int data;
+    private double data;
     private long time;
 
     public int getType() {
@@ -17,11 +17,11 @@ public class CollectData {
         this.type = type;
     }
 
-    public int getData() {
+    public double getData() {
         return data;
     }
 
-    public void setData(int data) {
+    public void setData(double data) {
         this.data = data;
     }
